+++ theme/boost_union/CHANGES.md
@@ -5,7 +5,13 @@ Changes
 -------
 
 ### Unreleased
+* 2023-01-06 - Improvement: Small language tweaks in self enrolment course banners
+* 2023-01-05 - Bugfix: Unparsable JSON in templates/core/full_header.mustache
 
+### v4.0-r9
+
+* 2022-12-31 - Feature: Add settings and layouts to enable/disable showing course images or a fallback image in the header of the course page, solves #77.
+* 2022-12-31 - Feature: Allow admins to define 'flavours' (i.e. special designs) which are applied to cohorts and / or course categories, solves #25.
 * 2022-12-19 - Feature: Allow admins to hide primary navigation items, solves #65.
 * 2022-12-14 - Feature: Built-in contact, help and maintenance pages, solves #150.
 * 2022-11-28 - Updated Moodle Plugin CI to latest upstream recommendations